Use the latest released version, not 'latest'
Prior to this, if you ran this script without specifying a version or
you specified 'latest', then the installer would be downloaded with
'latest' in the file name. This made it impossible to tell exactly which
version you just downloaded.

This commit will, if using 'latest', determine exactly which version
number that is and use that rather than the string, 'latest'.

That magic is done by looking at the last line of the page at
"http://versions.puppet.com.s3-website-us-west-2.amazonaws.com"
This isn't neccessarily a "stable" api, but it is what the Puppet
Enterprise download website uses, so hopefully it doesn't change.